Move embed_success to utils
